+++ assets/javascripts/kitten/components/buttons/button/test.js
@@ -56,10 +56,22 @@ describe('<Button />', () => {
     })
   })
 
-  describe('with `modifier` prop', () => {
+  describe('with `oxygen` `modifier` prop', () => {
     beforeEach(() => {
       component = renderer
         .create(<Button modifier="oxygen">MyButton</Button>)
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with `valid` `modifier` prop', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(<Button modifier="valid">MyButton</Button>)
         .toJSON()
     })